+++ lib/app.dart
@@ -5,11 +5,11 @@ import 'package:alex/internal/print.dart' as print;
 Future<int> run(List<String> args) async {
   try {
     return await AlexCommandRunner().run(args) ?? 2;
-  } on UsageException catch (e) {
-    print.exception(e);
+  } on UsageException catch (e, st) {
+    print.exception(e, st);
     return 64;
-  } catch (e) {
-    print.exception(e);
+  } catch (e, st) {
+    print.exception(e, st);
     return -1;
   }
 }