+++ lib/internal/print.dart
@@ -1,5 +1,6 @@
 import 'dart:io';
 
+import 'package:alex/src/const.dart';
 /// Prints some info message in output.
 void info(String message) => stdout.writeln(message);
 
@@ -7,7 +8,8 @@ void info(String message) => stdout.writeln(message);
 void error(String message) => stderr.writeln(message);
 
 /// Prints error message in error output.
-void exception(Object message) => error(message.toString());
+void exception(Object message, [StackTrace? stackTrace]) =>
+    error(message.toString() + (isDebug ? '\n$stackTrace' : ''));
 
 /// Prints some verbose message in output.
 // ignore: avoid_print