+++ lib/src/const.dart
@@ -1,0 +1,2 @@
+const bool isProduction = bool.fromEnvironment('dart.vm.product');
+const bool isDebug = !isProduction;