+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Prints stack trace in debug.
+
 ## 1.0.6 - 2022-02-07
 
 * [Finish Feature] Removes local feature branch if it's merged in remote.